requirements.txt
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
Deprecated==1.2.13
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dotenv==0.19.2
redis==4.1.4
rq==1.10.1
six==1.16.0
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
